python-integration-tests/test-requirements.txt
python-arango==7.3.4
pytest==7.1.2